🚚 [MOVE] : 컴포넌트 폴더 이동 및 common 폴더 삭제
