+++ src/components/Badge/EmojiBadge/PaperListEmojiBadge.jsx
