+++ src/components/Badge/EmojiBadge/PaperListEmojiBadge.style.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import { FONT_STYLE } from '../../../../style/fontStyle';
-import { COLORS } from '../../../../style/colorPalette';
+import { FONT_STYLE } from 'style/fontStyle';
+import { COLORS } from 'style/colorPalette';
 
 const BREAKPOINT_TABLET = 768;
 const BREAKPOINT_PC = 1200;
