+++ src/components/Badge/FromBadge/PaperListFromBadge.style.js
@@ -1,6 +1,6 @@
 import styled, { css } from 'styled-components';
-import { FONT_STYLE } from '../../../../style/fontStyle';
-import { COLORS } from '../../../../style/colorPalette';
+import { FONT_STYLE } from 'style/fontStyle';
+import { COLORS } from 'style/colorPalette';
 
 export const Container = styled.div`
   display: flex;
