--- src/components/Common/Badge/EmojiBadge/index.js
@@ -1,1 +1,0 @@
-export { default } from './PaperListEmojiBadge';
