--- src/components/Common/Badge/FromBadge/index.js
@@ -1,1 +1,0 @@
-export { default } from './PaperListFromBadge';
