+++ src/pages/PaperListPage/components/PaperCard/PaperCard.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import * as S from './PaperCard.style';
-import From from '../../../../components/Common/Badge/FromBadge';
-import Emoji from '../../../../components/Common/Badge/EmojiBadge';
-import PatternPurple from '../../../../assets/icon/pattern_purple.svg';
-import PatternBeige from '../../../../assets/icon/pattern_beige.svg';
-import PatternBlue from '../../../../assets/icon/pattern_blue.svg';
-import PatternGreen from '../../../../assets/icon/pattern_green.svg';
+import PaperListFromBadge from 'components/Badge/FromBadge/PaperListFromBadge';
+import PaperListEmojiBadge from 'components/Badge/EmojiBadge/PaperListEmojiBadge';
+import PatternPurple from 'assets/icon/pattern_purple.svg';
+import PatternBeige from 'assets/icon/pattern_beige.svg';
+import PatternBlue from 'assets/icon/pattern_blue.svg';
+import PatternGreen from 'assets/icon/pattern_green.svg';
 
 const BACKGROUND = {
   purple: { color: 'PURPLE_200', pattern: PatternPurple },
@@ -51,22 +51,22 @@ const PaperCard = ({ data = {}, slideIndex = 0 }) => {
           <S.Title $hasBackgroundImage={hasBackgroundImage}>
             {reduceText(name, 9)}
           </S.Title>
-          <From imgUrls={fromImgUrls} count={messageCount} />
+          <PaperListFromBadge imgUrls={fromImgUrls} count={messageCount} />
           <S.Description $hasBackgroundImage={hasBackgroundImage}>
             <S.Count>{messageCount}</S.Count>명이 작성했어요!
           </S.Description>
         </S.TextContainer>
         <S.Line />
-        <Emoji.Container>
+        <PaperListEmojiBadge.Container>
           {topReactions &&
             topReactions?.map((reaction) => (
-              <Emoji
+              <PaperListEmojiBadge
                 key={reaction.id}
                 emoji={reaction.emoji}
                 count={reaction.count}
               />
             ))}
-        </Emoji.Container>
+        </PaperListEmojiBadge.Container>
       </S.Wrapper>
       {!backgroundImageURL && (
         <S.Pattern src={BACKGROUND?.[backgroundColor]?.pattern} />
