+++ src/pages/PaperListPage/components/PaperCard/PaperCard.style.js
@@ -1,6 +1,6 @@
 import styled, { css } from 'styled-components';
-import { FONT_STYLE } from '../../../../style/fontStyle';
-import { COLORS } from '../../../../style/colorPalette';
+import { FONT_STYLE } from 'style/fontStyle';
+import { COLORS } from 'style/colorPalette';
 
 const BREAKPOINT_TABLET = 768;
 const BREAKPOINT_PC = 1200;
