--- src/pages/PaperListPage/components/PaperCard/index.js
@@ -1,1 +1,0 @@
-export { default } from './PaperCard.jsx';
