+++ src/pages/PostMessagePage/PostMessagePage.js
@@ -3,22 +3,22 @@ import * as S from './PostMessagePage.style.js';
 import arrowDownIcon from './asset/arrow_down.png';
 import arrowUpIcon from './asset/arrow_top.png';
 import TextEditor from './TextEditor';
-import { COLORS } from '../../style/colorPalette';
-import Header from '../../components/Common/Header/Header.jsx';
+import { COLORS } from 'style/colorPalette';
+import Header from 'components/Header/Header.jsx';
 import axios from 'axios';
 import { useParams, useNavigate } from 'react-router-dom';
-import { PurpleButton } from '../../components/Common/PurpleButton/PurpleButton.jsx';
-import sampleImg1 from '../../assets/images-message/1.png';
-import sampleImg2 from '../../assets/images-message/2.jpg';
-import sampleImg3 from '../../assets/images-message/3.jpg';
-import sampleImg4 from '../../assets/images-message/4.jpg';
-import sampleImg5 from '../../assets/images-message/5.jpg';
-import sampleImg6 from '../../assets/images-message/6.jpg';
-import sampleImg7 from '../../assets/images-message/7.jpg';
-import sampleImg8 from '../../assets/images-message/8.jpg';
-import sampleImg9 from '../../assets/images-message/9.jpg';
-import sampleImg10 from '../../assets/images-message/10.jpg';
-import { TextForm } from '../../components/Common/TextForm/TextForm.jsx';
+import { PurpleButton } from 'components/PurpleButton/PurpleButton.jsx';
+import sampleImg1 from 'assets/images-message/1.png';
+import sampleImg2 from 'assets/images-message/2.jpg';
+import sampleImg3 from 'assets/images-message/3.jpg';
+import sampleImg4 from 'assets/images-message/4.jpg';
+import sampleImg5 from 'assets/images-message/5.jpg';
+import sampleImg6 from 'assets/images-message/6.jpg';
+import sampleImg7 from 'assets/images-message/7.jpg';
+import sampleImg8 from 'assets/images-message/8.jpg';
+import sampleImg9 from 'assets/images-message/9.jpg';
+import sampleImg10 from 'assets/images-message/10.jpg';
+import { TextForm } from 'components/TextForm/TextForm.jsx';
 
 export const PostMessagePage = () => {
   const [isOpenRelation, setIsOpen] = useState(false);
