+++ src/pages/PostPage/components/PostImageButton/PostImageButton.js
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
 import * as S from './PostImageButton.style';
-import SelectImg from '../../assets/icon/background-selected.png';
-import sky from '../../assets/images/background-img1.jpeg';
-import book from '../../assets/images/background-img2.jpeg';
-import mountain from '../../assets/images/background-img3.jpeg';
-import ground from '../../assets/images/background-img4.jpeg';
+import SelectImg from 'assets/icon/background-selected.png';
+import sky from 'assets/images/background-img1.jpeg';
+import book from 'assets/images/background-img2.jpeg';
+import mountain from 'assets/images/background-img3.jpeg';
+import ground from 'assets/images/background-img4.jpeg';
 
 const INITIAL_IMAGE = {
   'https://picsum.photos/id/683/3840/2160': sky,
